+++ index.js
@@ -7,8 +7,7 @@ const Request = require("request");
 const supertest = require("supertest");
 const url = require("url");
 const vm = require("vm");
-const {Document, Fetch, Window, Compiler} = require("./lib");
-const {compile} = Compiler;
+const {Document, Fetch, Window} = require("./lib");
 const assert = require("assert");
 
 module.exports = Tallahassee;
@@ -75,18 +74,24 @@ function Tallahassee(app) {
   }
 
   function load(resp) {
-    let initialized, pending, currentPageXOffset, currentPageYOffset;
+    let pending, currentPageXOffset, currentPageYOffset;
     let elementsToScroll = () => {};
     const stickedElements = [];
 
-    compile();
-
     const headers = getHeaders(resp.request);
+    const document = Document(resp, agent.jar);
     const window = Window(resp, {
       fetch: Fetch(agent, resp),
+      get document() {
+        return document;
+      },
     });
-    const document = Document(resp, agent.jar);
-    window.document = document;
+
+    Object.defineProperty(document, "window", {
+      get() {
+        return window;
+      }
+    });
 
     const browserContext = {
       $: document.$,
@@ -119,13 +124,6 @@ function Tallahassee(app) {
     return browserContext;
 
     function focus() {
-      if (initialized) {
-        compile();
-      }
-
-      initialized = true;
-      global.window = window;
-      global.document = document;
     }
 
     function onDocumentSubmit(event) {
@@ -173,7 +171,7 @@ function Tallahassee(app) {
         if (scriptType && !/javascript/i.test(scriptType)) return;
 
         const scriptBody = $script.html();
-        if (scriptBody) vm.runInThisContext(scriptBody);
+        if (scriptBody) vm.runInNewContext(scriptBody, window);
       });
     }